refactor(views): centralize type definitions and constants for improved reusability and maintainability across views components
