+++ shared/src/containers/Views/hooks/useSaveViewFromCurrent.ts
@@ -1,5 +1,7 @@
 import { useCallback } from 'react'
-import { isViewStudioScope, ViewData, ViewSettings, ViewType } from '..'
+import type { ViewType } from '../types'
+import type { ViewData, ViewSettings } from '../context/ViewsContext'
+import { isViewStudioScope } from '../utils/isViewStudioScope'
 import { UseViewMutations } from './useViewsMutations'
 import { ViewListItemModel } from '@shared/api'
 import { toast } from 'react-toastify'
