+++ src/containers/Viewer/ViewerImage.tsx
@@ -2,7 +2,7 @@ import { FC, useRef, useState, useEffect, CSSProperties } from 'react'
 import { Image } from './Viewer.styled'
 import { useViewer } from '@context/ViewerContext'
 import styled, { keyframes } from 'styled-components'
-import { AnnotationsContainerDimensions } from './'
+import type { AnnotationsContainerDimensions } from './'
 import { Icon } from '@ynput/ayon-react-components'
 
 const spin = keyframes`
