+++ src/pages/ProjectListsPage/hooks/useListContextMenu.ts
@@ -1,11 +1,11 @@
 import { RowSelectionState } from '@tanstack/react-table'
-import { useListsContext } from '../context'
+import { useListsContext } from '../context/ListsContext'
 import { CommandEvent, useCreateContextMenu } from '@shared/containers/ContextMenu'
 import { useCallback } from 'react'
 import { useAppSelector } from '@state/store'
 import useClearListItems from './useClearListItems'
 import { useListsDataContext } from '../context/ListsDataContext'
-import { parseListFolderRowId } from '../util'
+import { parseListFolderRowId } from '../util/buildListsTableData'
 import { EntityListFolderModel } from '@shared/api'
 import { getPlatformShortcutKey, KeyMode, buildFolderHierarchy } from '@shared/util'
 import { usePowerpack, useProjectContext } from '@shared/context'
